+++ lib/src/connectivity/connectivity.dart
@@ -1,0 +1,2 @@
+export 'nano_connectivity.dart';
+export 'nano_connectivity_status.dart';