+++ lib/src/entity/entity.dart
@@ -1,0 +1,1 @@
+export 'nano_entity.dart';