+++ lib/src/equatable/equatable.dart
@@ -1,0 +1,1 @@
+export 'nano_equatable.dart';