+++ lib/src/plugins/plugins.dart
@@ -1,0 +1,1 @@
+export 'nano_core_plugin.dart';