+++ lib/src/storage/storage.dart
@@ -1,0 +1,1 @@
+export 'nano_storage.dart';