+++ lib/src/strategy/strategy.dart
@@ -1,0 +1,1 @@
+export 'nano_data_strategy.dart';